+++ Octokit.Reactive/Clients/ObservableRepositoryHooksClient.cs
@@ -84,7 +84,7 @@ namespace Octokit.Reactive
         /// <param name="name">The repository's name</param>
         /// <param name="hookId">The repository's hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#get-single-hook">API documentation</a> for more information.</remarks>
-        public IObservable<RepositoryHook> Get(string owner, string name, int hookId)
+        public IObservable<RepositoryHook> Get(string owner, string name, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -98,7 +98,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="hookId">The repository's hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#get-single-hook">API documentation</a> for more information.</remarks>
-        public IObservable<RepositoryHook> Get(long repositoryId, int hookId)
+        public IObservable<RepositoryHook> Get(long repositoryId, long hookId)
         {
             return _client.Get(repositoryId, hookId).ToObservable();
         }
@@ -140,7 +140,7 @@ namespace Octokit.Reactive
         /// <param name="hookId">The repository's hook id</param>
         /// <param name="hook">The requested changes to an edit repository hook</param>
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#edit-a-hook">API documentation</a> for more information.</remarks>
-        public IObservable<RepositoryHook> Edit(string owner, string name, int hookId, EditRepositoryHook hook)
+        public IObservable<RepositoryHook> Edit(string owner, string name, long hookId, EditRepositoryHook hook)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -156,7 +156,7 @@ namespace Octokit.Reactive
         /// <param name="hookId">The repository's hook id</param>
         /// <param name="hook">The requested changes to an edit repository hook</param>
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#edit-a-hook">API documentation</a> for more information.</remarks>
-        public IObservable<RepositoryHook> Edit(long repositoryId, int hookId, EditRepositoryHook hook)
+        public IObservable<RepositoryHook> Edit(long repositoryId, long hookId, EditRepositoryHook hook)
         {
             Ensure.ArgumentNotNull(hook, nameof(hook));
 
@@ -172,7 +172,7 @@ namespace Octokit.Reactive
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#test-a-hook">API documentation</a> for more information. 
         /// This will trigger the hook with the latest push to the current repository if the hook is subscribed to push events. If the hook 
         /// is not subscribed to push events, the server will respond with 204 but no test POST will be generated.</remarks>
-        public IObservable<Unit> Test(string owner, string name, int hookId)
+        public IObservable<Unit> Test(string owner, string name, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -188,7 +188,7 @@ namespace Octokit.Reactive
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#test-a-hook">API documentation</a> for more information. 
         /// This will trigger the hook with the latest push to the current repository if the hook is subscribed to push events. If the hook 
         /// is not subscribed to push events, the server will respond with 204 but no test POST will be generated.</remarks>
-        public IObservable<Unit> Test(long repositoryId, int hookId)
+        public IObservable<Unit> Test(long repositoryId, long hookId)
         {
             return _client.Test(repositoryId, hookId).ToObservable();
         }
@@ -200,7 +200,7 @@ namespace Octokit.Reactive
         /// <param name="name">The repository's name</param>
         /// <param name="hookId">The repository's hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#edit-a-hook">API documentation</a> for more information.</remarks>
-        public IObservable<Unit> Ping(string owner, string name, int hookId)
+        public IObservable<Unit> Ping(string owner, string name, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -214,7 +214,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="hookId">The repository's hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#edit-a-hook">API documentation</a> for more information.</remarks>
-        public IObservable<Unit> Ping(long repositoryId, int hookId)
+        public IObservable<Unit> Ping(long repositoryId, long hookId)
         {
             return _client.Ping(repositoryId, hookId).ToObservable();
         }
@@ -226,7 +226,7 @@ namespace Octokit.Reactive
         /// <param name="name">The repository's name</param>
         /// <param name="hookId">The repository's hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#delete-a-hook">API documentation</a> for more information.</remarks>
-        public IObservable<Unit> Delete(string owner, string name, int hookId)
+        public IObservable<Unit> Delete(string owner, string name, long hookId)
         {
             Ensure.ArgumentNotNullOrEmptyString(owner, nameof(owner));
             Ensure.ArgumentNotNullOrEmptyString(name, nameof(name));
@@ -240,7 +240,7 @@ namespace Octokit.Reactive
         /// <param name="repositoryId">The Id of the repository</param>
         /// <param name="hookId">The repository's hook id</param>
         /// <remarks>See <a href="http://developer.github.com/v3/repos/hooks/#delete-a-hook">API documentation</a> for more information.</remarks>
-        public IObservable<Unit> Delete(long repositoryId, int hookId)
+        public IObservable<Unit> Delete(long repositoryId, long hookId)
         {
             return _client.Delete(repositoryId, hookId).ToObservable();
         }
